+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Default search engine changed to geminispace.info from gus.guru
 - The user's terminal theme colors are used by default (#181)
 - By default, non-gemini URI schemes are opened in the default application. This requires a config change for previous users, see the [wiki](https://github.com/makeworld-the-better-one/amfora/wiki/Handling-Other-URL-Schemes) (#207)
+- Windows uses paths set by `XDG` variables over `APPDATA` if they are set (#255)
 
 ## Removed
 - Favicon support (#199)